Implement user separation for armor pieces
- Remove manual user ID field from add armor piece form and servlet
- Filter view to show only the logged-in user's armor pieces
- Change AppUser userId field from Long to int for consistency

+++ src/main/java/edu/matc/inventory/entity/AppUser.java
@@ -16,7 +16,7 @@ public class AppUser {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private Long userId;
+    private int userId;
 
     @Column(name = "auth_subject")
     private String authSubject;
@@ -39,11 +39,11 @@ public class AppUser {
         this.displayName = displayName;
     }
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
